refactor: Code improvements
